+++ assets/styles/baseInputStyle.js
@@ -1,0 +1,6 @@
+export const baseInputStyle = {
+borderRadius: "5px !important",
+"&:focus": {
+    borderColor: "#9CA5F8",
+}
+}
